fix(native): bound client connect and split lifecycle modules
