+++ rust/Cargo.toml
@@ -22,7 +22,7 @@ rust-version = "1.88"
 hex = "0.4.3"
 getrandom = "0.3.3"
 libc = "0.2.175"
-nix = { version = "0.31.3", features = ["socket"] }
+nix = { version = "0.31.3", features = ["fs", "poll", "socket"] }
 regex = "1.11.2"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.143"
