+++ rust/crates/guard-runtime/src/managed_resident_tests.rs
@@ -1,0 +1,17 @@
+use super::*;
+
+#[test]
+fn generation_parser_rejects_zero_and_non_numeric() {
+    assert!(parse_generation("0").is_err());
+    assert!(parse_generation("not-a-number").is_err());
+    assert_eq!(parse_generation("7").unwrap(), 7);
+}
+
+#[test]
+fn client_deadline_is_bounded() {
+    assert_eq!(
+        client_timeout(br#"{"deadline_budget_ms":999999}"#),
+        Duration::from_secs(9)
+    );
+    assert_eq!(client_timeout(br#"{}"#), Duration::from_millis(750));
+}
