+++ tests/test_rust_pretool_authority.py
@@ -434,3 +434,27 @@ def test_hook_worker_fails_closed_for_non_command_pretool_without_native_result(
     )
     assert result["decision"] == "deny"
     assert result["reason_code"] == "native_pre_tool_unavailable"
+
+
+def test_hook_worker_leaves_out_of_scope_events_to_existing_handling(
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    monkeypatch.setattr(
+        "codex_plugin_scanner.guard.daemon.hook_worker.native_mode",
+        lambda: "auto",
+    )
+    monkeypatch.setattr(
+        "codex_plugin_scanner.guard.daemon.hook_worker.review_raw_hook_native",
+        lambda *_args, **_kwargs: pytest.fail("out-of-scope event reached the native edge"),
+    )
+    worker = HookWorker(store=GuardStore(tmp_path / "guard-home"))
+    with pytest.raises(HookWorkerUnsupported, match="fast path supports"):
+        worker.review_http_payload(
+            payload={"hook_event_name": "PermissionRequest", "tool_input": {"command": "pwd"}},
+            params={},
+            default_harness="claude-code",
+            home_dir=tmp_path / "home",
+            guard_home=tmp_path / "guard-home",
+            workspace=tmp_path / "workspace",
+        )
